type switch

+++ src/shapes.c
@@ -111,7 +111,7 @@ void S2D_DrawQuad(GLfloat x1, GLfloat y1,
 /*
  * Draw a rectangle, using two triangles. Simplified version of quad
  */
-void S2D_DrawRect_XYWH(GLfloat x, GLfloat y, int width, int height, S2D_Color color, bool filled) {
+void S2D_DrawRect_XYWH(GLfloat x, GLfloat y, GLfloat width, GLfloat height, S2D_Color color, bool filled) {
   GLfloat x1, y1, r1, g1, b1, a1,
           x2, y2, r2, g2, b2, a2,
           x3, y3, r3, g3, b3, a3,
